requirements.txt
coverage~=7.2.0
django~=4.1.7
django-filter~=22.1
django-model-utils~=4.3.1
djangorestframework~=3.14.0
environs~=9.5.0
pytest~=7.2.1
pytest-django~=4.5.2
python-dotenv~=0.21.1
requests~=2.28.2
pytest-cov~=4.0.0
pytest-mock~=3.10.0